+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.jsx
@@ -60,20 +60,15 @@ class RicosTestApp extends PureComponent {
     const { isMobile, contentState, locale, seoMode, testAppConfig } = this.props;
 
     return (
-      <>
-        <RicosViewer
-          plugins={viewerPlugins(testAppConfig.plugins)}
-          content={contentState}
-          isMobile={isMobile}
-          locale={locale}
-          cssOverride={theme}
-          seoSettings={seoMode}
-          preview={testAppConfig.showDefaultPreview && previewSettings()}
-        />
-        <TextSelectionToolbar container={this.viewerRef.current}>
-          {selectedText => <TwitterButton selectedText={selectedText} />}
-        </TextSelectionToolbar>
-      </>
+      <RicosViewer
+        plugins={viewerPlugins(testAppConfig.plugins)}
+        content={contentState}
+        isMobile={isMobile}
+        locale={locale}
+        cssOverride={theme}
+        seoSettings={seoMode}
+        preview={testAppConfig.showDefaultPreview && previewSettings()}
+      />
     );
   };
 
@@ -96,6 +91,9 @@ class RicosTestApp extends PureComponent {
             ref={this.viewerRef}
           >
             {this.renderViewer()}
+            <TextSelectionToolbar container={this.viewerRef.current}>
+              {selectedText => <TwitterButton selectedText={selectedText} />}
+            </TextSelectionToolbar>
           </div>
         </div>
       </div>